Extract runner logic from JavaGeneratorMain into Runnable interface
